NO-SNOW: Restructure python wrapper dir
