+++ python/.gitignore
@@ -1,0 +1,114 @@
+reports/
+
+# Byte-compiled / optimized / DLL files
+.mypy_cache
+__pycache__
+*.py[cod]
+*$py.class
+
+# C extensions
+*.a
+*.dll
+*.exe
+*.o
+*.so
+*.dylib
+
+# Distribution / packaging
+.Python
+env/
+build/
+develop-eggs/
+dist/
+downloads/
+eggs/
+.eggs/
+lib/
+lib64/
+parts/
+sdist/
+var/
+*.egg-info/
+.installed.cfg
+*.egg
+
+# PyInstaller
+#  Usually these files are written by a python script from a template
+#  before PyInstaller builds the exe, so as to inject date/other infos into it.
+*.manifest
+*.spec
+
+# Installer logs
+pip-log.txt
+pip-delete-this-directory.txt
+
+# Unit test / coverage reports
+htmlcov/
+.tox/
+.coverage
+.coverage.*
+.cache
+nosetests.xml
+coverage.xml
+*,cover
+.hypothesis/
+
+# Translations
+*.mo
+*.pot
+
+# Sphinx documentation
+docs/_build/
+
+# PyBuilder
+target/
+
+# IPython Notebook
+.ipynb_checkpoints
+
+# pyenv
+.python-version
+
+# celery beat schedule file
+celerybeat-schedule
+
+# dotenv
+.env
+
+# virtualenv
+venv*/
+ENV/
+
+# Spyder project settings
+.spyderproject
+
+# Rope project settings
+.ropeproject
+
+# vim
+*.swp
+
+# others
+generated_version.py
+*coverage.xml
+.pytest_cache/
+*.egg-info/
+.tox/
+
+# WhiteSource Scan
+wss*.config
+wss-unified-agent.jar
+whitesource/
+
+# core dumps
+core.*
+
+# Debug specific
+.gdb_history
+
+# Editor specific
+.idea/
+*iml
+.vscode/
+*.code-workspace
+
