+++ python/src/snowflake/ud_connector/_internal/protobuf_gen/proto_exception.py
