+++ python/src/snowflake/ud_connector/cursor.py
@@ -5,7 +5,7 @@ This module defines the Cursor class as specified in PEP 249.
 """
 from .exceptions import NotSupportedError
 import pyarrow
-from .protobuf_gen.database_driver_v1_pb2 import *
+
 
 class Cursor:
     """
