+++ python/src/snowflake/ud_connector/exceptions.py
