+++ python/tests/compatibility.py
