+++ python/tests/connector_types.py
