+++ python/tests/integ/utils_put_get.py
@@ -3,7 +3,7 @@ import pytest
 import uuid
 import zstandard as zstd
 from pathlib import Path
-from pep249_dbapi.cursor import Cursor
+from python.cursor import Cursor
 
 GET_ROW_FILE_IDX = 0
 GET_ROW_SIZE_IDX = 1
