+++ python/tox.ini
@@ -17,11 +17,11 @@ commands = uv run pytest tests -q {posargs}
 
 [testenv:lint]
 deps = .[dev]
-commands = uv run flake8 pep249_dbapi
+commands = uv run flake8 src/snowflake
 
 [testenv:type]
 deps = .[dev]
-commands = uv run mypy pep249_dbapi
+commands = uv run mypy src/snowflake
 
 [testenv:py{39,310,311,312,313}-unit]
 commands = uv run pytest tests/unit {posargs}
